pyproject.toml
[project]
name = "video-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python",
    "numpy",
    "ruff",
    "confluent-kafka>=2.8.2",
    "flask>=3.1.0",
    "kafka-python>=2.0.5",
    "nats-py>=2.9.0",
]

[tool.ruff]
ignore = ["E501"]  # Ignore line length errors
select = ["E", "F", "W"]  # Select specific error codes
per-file-ignores = {}

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
known_third_party = "ruff"
known_first_party = "video-analytics"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = "setup.py"
